refactor: FuelNumberGenerator, GameView 리팩터링
 - 기존 NumberSetGenerator에서 FuelNumberGenerator로 클래스 이름 변경
 - 그에 따른 관련 클래스 리팩터링
 - GameView 클래스 print할 문구 수정
