+++ src/main/java/kr/insup/utils/FuelNumberGenerator.java
@@ -1,27 +1,20 @@
-package kr.insup.domain;
+package kr.insup.utils;
 
 import kr.insup.utils.RandomValueGenerator;
 
-public class NumberSetGenerator {
-
-    private final int carNumber;
-
-    public NumberSetGenerator(int carNumber) {
-        this.carNumber = carNumber;
-    }
+public class FuelNumberGenerator {
 
     /**
      * 차의 수만큼 랜덤값 배열을 생성한다.
      *
+     * @param carNumber 차의 수
      * @return 차의 수만큼의 크기를 가진 랜덤값 배열
      */
-    public int[] makeOneSet() {
+    public static int[] makeFuelNumbers(int carNumber) {
         int[] numberSet = new int[carNumber];
         for (int i = 0; i < carNumber; i++) {
             numberSet[i] = RandomValueGenerator.generateRandomValue();
         }
         return numberSet;
     }
-
-
 }
